Rename $kind Microsoft.SkillDialog => Microsoft.BeginSkillDialog Rename class AdaptiveSkillDialog => BeginSkillDialog Move files into actions folders replace all $kind => new $kind
