+++ libraries/Microsoft.Bot.Builder.Dialogs.Adaptive/Actions/BeginSkillDialog.cs
@@ -12,15 +12,18 @@ using Microsoft.Bot.Builder.Skills;
 using Microsoft.Bot.Schema;
 using Newtonsoft.Json;
 
-namespace Microsoft.Bot.Builder.Dialogs.Adaptive.Skills
+namespace Microsoft.Bot.Builder.Dialogs.Adaptive.Actions
 {
-    public class AdaptiveSkillDialog : SkillDialog
+    /// <summary>
+    /// Begin a SkillDialog.
+    /// </summary>
+    public class BeginSkillDialog : SkillDialog
     {
         [JsonProperty("$kind")]
-        public const string Kind = "Microsoft.SkillDialog";
+        public const string Kind = "Microsoft.BeginSkillDialog";
 
         [JsonConstructor]
-        public AdaptiveSkillDialog([CallerFilePath] string callerPath = "", [CallerLineNumber] int callerLine = 0)
+        public BeginSkillDialog([CallerFilePath] string callerPath = "", [CallerLineNumber] int callerLine = 0)
             : base(new SkillDialogOptions())
         {
             DialogOptions.Skill = new BotFrameworkSkill();
