+++ libraries/Microsoft.Bot.Builder.Dialogs.Adaptive/AdaptiveComponentRegistration.cs
@@ -13,7 +13,6 @@ using Microsoft.Bot.Builder.Dialogs.Adaptive.Generators;
 using Microsoft.Bot.Builder.Dialogs.Adaptive.Input;
 using Microsoft.Bot.Builder.Dialogs.Adaptive.Recognizers;
 using Microsoft.Bot.Builder.Dialogs.Adaptive.Selectors;
-using Microsoft.Bot.Builder.Dialogs.Adaptive.Skills;
 using Microsoft.Bot.Builder.Dialogs.Choices;
 using Microsoft.Bot.Builder.Dialogs.Debugging;
 using Microsoft.Bot.Builder.Dialogs.Declarative;
@@ -59,6 +58,7 @@ namespace Microsoft.Bot.Builder.Dialogs.Adaptive
 
             // Actions
             yield return new DeclarativeType<BeginDialog>(BeginDialog.Kind);
+            yield return new DeclarativeType<BeginSkillDialog>(BeginSkillDialog.Kind);
             yield return new DeclarativeType<CancelDialog>(CancelDialog.Kind);
             yield return new DeclarativeType<CancelAllDialogs>(CancelAllDialogs.Kind);
             yield return new DeclarativeType<DebugBreak>(DebugBreak.Kind);
@@ -141,7 +141,6 @@ namespace Microsoft.Bot.Builder.Dialogs.Adaptive
 
             // Dialogs
             yield return new DeclarativeType<AdaptiveDialog>(AdaptiveDialog.Kind);
-            yield return new DeclarativeType<AdaptiveSkillDialog>(AdaptiveSkillDialog.Kind);
 
             // register x.dialog.schema/x.dialog as DynamicBeginDialog $kind="x" => DynamicBeginDialog(x.dialog) resource.
             foreach (var schema in resourceExplorer.GetResources(".schema").Where(s => resourceExplorer.GetTypeForKind(Path.GetFileNameWithoutExtension(s.Id)) == null))
